# swagger.yaml
openapi: 3.0.3
info:
  title: Ibrat Lead Calling System API
  description: |
    Complete backend API for AI-powered lead calling system with admin and operator panels
    
    ## 🚀 Quick Start Guide
    
    ### **1. Authentication Flow**
    1. **Register User**: `POST /api/auth/register` (admin only)
       - Use test credentials: `john.doe@ibrat.com` / `ibrat123`
    2. **Login**: `POST /api/auth/login` with same credentials
       - Copy JWT token from response
    3. **Use APIs**: Add `Authorization: Bearer <token>` header
    
    ### **2. Test Credentials**
    - **Admin**: `john.doe@ibrat.com` / `ibrat123`
    - **Salesperson**: `jane.smith@ibrat.com` / `ibrat123`
    
    ### **3. API Access**
    - **Admin Panel**: Full system access
    - **Operator Panel**: Work interface only
    - **Real-time**: WebSocket connections
    
    ### **4. Swagger Features**
    - Click "Try it out" to test APIs
    - Use "Authorize" button for JWT tokens
    - View request/response schemas
  version: 1.0.0
  contact:
    name: Backend Team
    email: backend@ibrat.com
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.ibrat.com
    description: Production server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token from /api/auth/login
        
        **How to use:**
        1. Register: `POST /api/auth/register`
        2. Login: `POST /api/auth/login` 
        3. Copy token from response
        4. Click "Authorize" button above
        5. Enter: `Bearer <your_token>`
        
        **Example:**
        ```
        Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjUwN2YxZjc3YmNmODZjZDc5OTQzOTAxMSIsIm5hbWUiOiJKb2huIERvZSIsImVtYWlsIjoiam9obi5kb2VAaWJyYXQuY29tIiwicm9sZSI6ImFkbWluIiwiaWF0IjoxNzM1NzE5NjAwLCJleHAiOjE3MzU3MjU2MDB9.example
        ```
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439011"
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        phone:
          type: string
          example: "+1234567890"
        role:
          type: string
          enum: [admin, salesperson]
          example: "salesperson"
        isActive:
          type: boolean
          example: true
        isAvailable:
          type: boolean
          example: true
        lastLogin:
          type: string
          format: date-time
        callStats:
          type: object
          properties:
            totalCalls:
              type: number
              example: 150
            successfulCalls:
              type: number
              example: 120
            totalDuration:
              type: number
              example: 7200
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "john.doe@ibrat.com"
        password:
          type: string
          example: "ibrat123"
    RegisterRequest:
      type: object
      required:
        - name
        - email
        - phone
        - password
        - role
      properties:
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john.doe@ibrat.com"
        phone:
          type: string
          example: "+1234567890"
        password:
          type: string
          minLength: 6
          example: "ibrat123"
        role:
          type: string
          enum: [admin, salesperson]
          example: "admin"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Validation error"
        message:
          type: string
          example: "Phone number is required"
paths:
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Register a new user
      description: |
        Create a new user account (admin only)
        
        **Example Workflow:**
        1. Register user with these credentials
        2. Use same credentials to login
        3. Get JWT token for API access
        
        **Test Credentials:**
        - Email: john.doe@ibrat.com
        - Password: ibrat123
        - Role: admin
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              admin_user:
                summary: Admin User Registration
                description: Register an admin user with test credentials
                value:
                  name: "John Doe"
                  email: "john.doe@ibrat.com"
                  phone: "+1234567890"
                  password: "ibrat123"
                  role: "admin"
              salesperson_user:
                summary: Salesperson User Registration
                description: Register a salesperson user
                value:
                  name: "Jane Smith"
                  email: "jane.smith@ibrat.com"
                  phone: "+1234567891"
                  password: "ibrat123"
                  role: "salesperson"
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User registered successfully"
                  user:
                    type: object
                    properties:
                      _id:
                        type: string
                        example: "507f1f77bcf86cd799439011"
                      name:
                        type: string
                        example: "John Doe"
                      email:
                        type: string
                        example: "john.doe@ibrat.com"
                      phone:
                        type: string
                        example: "+1234567890"
                      role:
                        type: string
                        example: "admin"
                      isActive:
                        type: boolean
                        example: true
                      isAvailable:
                        type: boolean
                        example: true
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: |
        Authenticate user and get JWT token
        
        **Example Workflow:**
        1. Use credentials from registration
        2. Get JWT token for API access
        3. Include token in Authorization header
        
        **Test Credentials:**
        - Email: john.doe@ibrat.com
        - Password: ibrat123
        
        **Usage:**
        - Copy JWT token from response
        - Add to Authorization header: `Bearer <token>`
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              admin_login:
                summary: Admin Login
                description: Login with admin credentials
                value:
                  email: "john.doe@ibrat.com"
                  password: "ibrat123"
              salesperson_login:
                summary: Salesperson Login
                description: Login with salesperson credentials
                value:
                  email: "jane.smith@ibrat.com"
                  password: "ibrat123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Login successful"
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjUwN2YxZjc3YmNmODZjZDc5OTQzOTAxMSIsIm5hbWUiOiJKb2huIERvZSIsImVtYWlsIjoiam9obi5kb2VAaWJyYXQuY29tIiwicm9sZSI6ImFkbWluIiwiaWF0IjoxNzM1NzE5NjAwLCJleHAiOjE3MzU3MjU2MDB9.example"
                  user:
                    type: object
                    properties:
                      _id:
                        type: string
                        example: "507f1f77bcf86cd799439011"
                      name:
                        type: string
                        example: "John Doe"
                      email:
                        type: string
                        example: "john.doe@ibrat.com"
                      phone:
                        type: string
                        example: "+1234567890"
                      role:
                        type: string
                        example: "admin"
                      isActive:
                        type: boolean
                        example: true
                      isAvailable:
                        type: boolean
                        example: true
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/me:
    get:
      tags: [Authentication]
      summary: Get current user profile
      description: Retrieve current authenticated user's profile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
...